requirements.txt
numpy
scipy
pyyaml
matplotlib
plotly
numpy-stl
pandas
